Move ENTITY_CATEGORIES into homeassistant.helpers.entity
